projects/capstone-rag-qa-system/requirements.txt
# DocuMind AI - 依赖列表

# ============================================
# Web Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
streamlit>=1.31.0
httpx>=0.26.0

# ============================================
# LLM & Embedding
# ============================================
transformers>=4.37.0
sentence-transformers>=2.3.0
torch>=2.1.0
accelerate>=0.26.0
bitsandbytes>=0.42.0

# ============================================
# Vector Store
# ============================================
faiss-cpu>=1.7.4

# ============================================
# Document Parsing
# ============================================
PyMuPDF>=1.23.0
python-docx>=1.1.0
markdown>=3.5.0

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ============================================
# Utils
# ============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ============================================
# Development
# ============================================
black>=23.12.0
isort>=5.13.0
